pyproject.toml
[tool.poetry]
name = "myfl"
version = "0.1.0"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
typer = "^0.4.1"
# mpi4py = "^3.1.3"
pydantic = "^1.8.2"
websocket-client = "^1.3.2"
httpx = "^0.22.0"
fastapi = "^0.76.0"
starlette = "^0.18.0"

[tool.poetry.dev-dependencies]
# formatter
black = "^22.3.0"
isort = "^5.8.0"
flake8 = "^3.9.0"
mypy = "^0.812"

# test
pytest = "^6.2.2"
pytest-cov = "^3.0.0"
pre-commit = "^2.12.0"

# mkdocs
mkdocs = "^1.2.2"
mkautodoc = "^0.1.0"
mkdocs-material = "^7.2.2"
pymdown-extensions = "^8.2"
mkdocs-print-site-plugin = "^1.2.3"
ipykernel = "^6.13.0"
tensorflow = "^2.8.0"
torchvision = "^0.12.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
matplotlib = "^3.5.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
